Clang (macos) build fixes post-2.10.0 merge

+++ video/vdu_sys.h
@@ -382,7 +382,7 @@ void VDUStreamProcessor::sendCursorPosition() {
 //
 void VDUStreamProcessor::sendScreenChar(char c) {
 	uint8_t packet[] = {
-		c,
+		(uint8_t)c,
 	};
 	send_packet(PACKET_SCRCHAR, sizeof packet, packet);
 }
